# Improve DLN proof verification performance for large signing groups